refactor: Update import paths for github package
