+++ metrics/internal/test/cmd.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/mozilla-services/rapid-release-model/metrics/internal/export"
 	"github.com/mozilla-services/rapid-release-model/metrics/internal/factory"
-	"github.com/mozilla-services/rapid-release-model/metrics/internal/github"
 	"github.com/mozilla-services/rapid-release-model/metrics/internal/grafana"
+	"github.com/mozilla-services/rapid-release-model/pkg/github"
 	"github.com/spf13/cobra"
 )
